Finish compatibility transformations
Fix test & move the ugly logic into its own module
